11/26 starpoint type change

+++ sample/src/main/java/zebra/beans/MyReviewItem.java
@@ -7,10 +7,10 @@ public class MyReviewItem {
 
     public String productImage;
     public String memberEmail;
-    public double ratingBar;
+    public int ratingBar;
     public String reviewText;
 
-    public MyReviewItem(String productImage, String memberEmail, double ratingBar, String reviewText){
+    public MyReviewItem(String productImage, String memberEmail, int ratingBar, String reviewText){
         this.productImage = productImage;
         this.memberEmail = memberEmail;
         this.ratingBar = ratingBar;
