+++ sample/src/main/java/zebra/fragments/CategorySearchFragment.java
@@ -66,7 +66,7 @@ public class CategorySearchFragment extends Fragment {
 
         String productUrl;
         String productName;
-        double starPoint;
+        int starPoint;
         String barcode;
 
         for (int i = 0; i < result.productInfo.size(); i++) {
