+++ sample/src/main/java/zebra/views/MyReviewItemView.java
@@ -48,7 +48,7 @@ public class MyReviewItemView extends FrameLayout {
             }
         });
         memberEmail.setText(item.memberEmail);
-        ratingBar.setRating((float)item.ratingBar);
+        ratingBar.setRating((int)item.ratingBar);
         reviewText.setText(item.reviewText);
     }
