+++ sample/src/main/java/zebra/views/SearchItemView.java
@@ -36,6 +36,6 @@ public class SearchItemView extends FrameLayout {
     public void setViewItem(SearchItem item) {
         Glide.with(getContext()).load(item.productUrl).into(productImage);
         productName.setText(item.productName);
-        ratingBar.setRating((float)item.ratingBar);
+        ratingBar.setRating((int)item.ratingBar);
     }
 }
